Refactor generateAnnotations to lib/diff.js
