+++ lib/actions.js
@@ -1,8 +1,8 @@
 const logger = require('./logger')
 
 const formatFile = require('./format')
-const { getStylefile, getPRFileList, getFile, generateAnnotations }
-  = require('./hub')
+const { getStylefile, getPRFileList, getFile } = require('./hub')
+const { generateAnnotations } = require('./diff')
 
 async function asyncForEach(array, callback) {
   for (let index = 0; index < array.length; index++) {
@@ -169,6 +169,7 @@ async function lint(
 
     // Format file
     const transformed = await formatFile(file, style)
+
     if (!transformed) {
       error(`Error transforming ${filename}`)
       skipped_filenames.push(filename)
@@ -179,7 +180,7 @@ async function lint(
     }
 
     // Generate annotations
-    const file_annotations = generateAnnotations(transformed, file)
+    const file_annotations = await generateAnnotations(transformed, file)
 
     touched_lines = Number(file_annotations.lines) + Number(touched_lines)
     annotations = annotations.concat(file_annotations.annotations)
