+++ lib/hub.js
@@ -1,5 +1,4 @@
 const yaml = require('js-yaml')
-const { structuredPatch } = require('diff')
 
 async function getStylefile({owner, repo, ref}, repos, info) {
   const stylefileName = '.clang-format'
@@ -100,32 +99,8 @@ async function getFile (git, {owner, repo, filename, sha}) {
     })
 }
 
-function generateAnnotations({ filename, content }, original) {
-  const annotations = []
-  let touchedLines = 0
-
-  const diff = structuredPatch(filename, filename, original.content, content)
-  diff.hunks.forEach(({oldStart, oldLines}) => {
-    const annotation = {
-      path: filename,
-      start_line: oldStart,
-      end_line: oldStart + oldLines,
-      annotation_level: 'failure',
-      message: `Lines ${oldStart}-${oldStart+oldLines} need formatting.`,
-    }
-    annotations.push(annotation)
-    touchedLines += oldLines
-  })
-
-  return {
-    annotations,
-    lines: touchedLines,
-  }
-}
-
 module.exports = {
   getStylefile,
   getPRFileList,
   getFile,
-  generateAnnotations,
 }
